+++ Assets/VRM/UniVRM/Editor/Format/VRMExporterWizard.cs.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: a1429b9028f33544e94aa367c2acb7fb
+guid: a8e41aa30fcc76e43ad8588aef8572ea
+timeCreated: 1520491195
+licenseType: Free
 MonoImporter:
-  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
